requirements.txt
groq==0.4.1
sentence-transformers==3.0.0
faiss-cpu==1.7.4
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.3
tiktoken==0.5.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
